Minimised code
Removed reliance on LayoutPanels. I now simply populate the form with text-boxes. The construction relies fully on the iterator and the fact that it can be factored by multiples of 3 every cell line and by multiples of 9 at end of every cell line and by multiples of 27 at every major line end.