Initial commit of SlackBot
Create a SlackBot in Glitch using Bolt. It responds to direct mentions and if a user types in an email address, it will echo the email address back to the user. This is intended to fetch some info from a repo based on the typed email address string. Will add this feature later.